// src/common/claims/claim.enum.ts
export enum Claim {
    USER_CREATE = 'user.create',
    USER_VIEW = 'user.view',
    USER_EDIT = 'user.edit',
    USER_DELETE = 'user.delete',
    USER_UPLOAD = 'user.upload',
    USER_DOWNLOAD = 'user.download',

    INVOICE_TPP_CREATE = 'invoice-tpp.create',
    INVOICE_TPP_VIEW = 'invoice-tpp.view',
    INVOICE_TPP_EDIT = 'invoice-tpp.edit',
    INVOICE_TPP_DELETE = 'invoice-tpp.delete',
    INVOICE_TPP_UPLOAD = 'invoice-tpp.upload',
    INVOICE_TPP_DOWNLOAD = 'invoice-tpp.download',

    INVOICE_LFI_CREATE = 'invoice-lfi.create',
    INVOICE_LFI_VIEW = 'invoice-lfi.view',
    INVOICE_LFI_EDIT = 'invoice-lfi.edit',
    INVOICE_LFI_DELETE = 'invoice-lfi.delete',
    INVOICE_LFI_UPLOAD = 'invoice-lfi.upload',
    INVOICE_LFI_DOWNLOAD = 'invoice-lfi.download',  

    BILLING_TPP_CREATE = 'billing-tpp.create',
    BILLING_TPP_VIEW = 'billing-tpp.view',
    BILLING_TPP_EDIT = 'billing-tpp.edit',
    BILLING_TPP_DELETE = 'billing-tpp.delete',
    BILLING_TPP_UPLOAD = 'billing-tpp.upload',
    BILLING_TPP_DOWNLOAD = 'billing-tpp.download',

    BILLING_LFI_CREATE = 'billing-lfi.create',
    BILLING_LFI_VIEW = 'billing-lfi.view',
    BILLING_LFI_EDIT = 'billing-lfi.edit',
    BILLING_LFI_DELETE = 'billing-lfi.delete',
    BILLING_LFI_UPLOAD = 'billing-lfi.upload',
    BILLING_LFI_DOWNLOAD = 'billing-lfi.download',   

    GLOBAL_CONFIGURATION_CREATE = 'global-configuration.create',
    GLOBAL_CONFIGURATION_VIEW = 'global-configuration.view',
    GLOBAL_CONFIGURATION_EDIT = 'global-configuration.edit',
    GLOBAL_CONFIGURATION_DELETE = 'global-configuration.delete', 
    GLOBAL_CONFIGURATION_UPLOAD = 'global-configuration.upload',
    GLOBAL_CONFIGURATION_DOWNLOAD = 'global-configuration.download',  

    DASHBOARD_CREATE = 'dashboard.create',
    DASHBOARD_VIEW = 'dashboard.view',
    DASHBOARD_EDIT = 'dashboard.edit',
    DASHBOARD_DELETE = 'dashboard.delete',
    DASHBOARD_UPLOAD = 'dashboard.upload',
    DASHBOARD_DOWNLOAD = 'dashboard.download',
    
    LOG_CREATE = 'log.create',
    LOG_VIEW = 'log.view',
    LOG_EDIT = 'log.edit',
    LOG_DELETE = 'log.delete',
    LOG_UPLOAD = 'log.upload',
    LOG_DOWNLOAD = 'log.download',  

    TPP_CONFIGURATION_CREATE = 'tpp-configuration.create',
    TPP_CONFIGURATION_VIEW = 'tpp-configuration.view',
    TPP_CONFIGURATION_EDIT = 'tpp-configuration.edit',
    TPP_CONFIGURATION_DELETE = 'tpp-configuration.delete',
    TPP_CONFIGURATION_UPLOAD = 'tpp-configuration.upload',
    TPP_CONFIGURATION_DOWNLOAD = 'tpp-configuration.download', 

    LFI_DIRECTORY_CREATE = 'lfi-directory.create',
    LFI_DIRECTORY_VIEW = 'lfi-directory.view',
    LFI_DIRECTORY_EDIT = 'lfi-directory.edit',
    LFI_DIRECTORY_DELETE = 'lfi-directory.delete',
    LFI_DIRECTORY_UPLOAD = 'lfi-directory.upload',
    LFI_DIRECTORY_DOWNLOAD = 'lfi-directory.download',  

    DATA_UPLOADER_CREATE = 'data-uploader.create',
    DATA_UPLOADER_VIEW = 'data-uploader.view',
    DATA_UPLOADER_EDIT = 'data-uploader.edit',
    DATA_UPLOADER_DELETE = 'data-uploader.delete',
    DATA_UPLOADER_UPLOAD = 'data-uploader.upload',
    DATA_UPLOADER_DOWNLOAD = 'data-uploader.download',  
}